use get-chart-data to get lab results
