+++ packages/utils/lib/helpers/visit-note/progress-note-chart-data-requested-fields.helper.ts
@@ -18,6 +18,7 @@ export const getProgressNoteChartDataRequestedFields = (): ChartDataRequestedFie
       .map((name) => (createVitalsSearchConfig(name, 'encounter').searchParams as { _tag: string })._tag)
       .join(','),
   },
+  labResults: {},
 });
 
 export const telemedProgressNoteChartDataRequestedFields: ChartDataRequestedFields = {
