+++ packages/utils/lib/types/api/chart-data/chart-data.types.ts
@@ -34,6 +34,7 @@ import {
   InPersonExamCardsNames,
   InPersonExamFieldsNames,
 } from './save-chart-data.types';
+import { EncounterLabResult } from '../lab';
 
 export interface ChartDataFields {
   chiefComplaint?: FreeTextNoteDTO;
@@ -59,6 +60,7 @@ export interface ChartDataFields {
   notes?: NoteDTO[];
   vitalsObservations?: VitalsObservationDTO[];
   birthHistory?: BirthHistoryDTO[];
+  labResults?: EncounterLabResult;
 }
 
 export type ChartDataFieldsKeys = keyof ChartDataFields;
